chore: Use RiskKey
